Pairwise Distinct Summands
Task. 
____

The goal of this problem is to represent a given positive integer n as a sum of as many pairwise
distinct positive integers as possible. That is, to find the maximum k such that n can be written as
a1 + a2 + · · · + ak where a1, . . . , ak are positive integers and ai !=  aj for all 1 ≤ i < j ≤ k.

Input Format. 
___________
The input consists of a single integer n.